Added algo setup
